refactor: decouple client elapsed seconds
