+++ hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useClientElapsedSeconds';
 export * from './useDominantColor';
 export * from './useFullScreen';
 export * from './useNowPlaying';
